Refactor/Bootcamp changes
