+++ dxc/ai/global_variables/globals_file.py
